+++ examples/async_demo.py
@@ -6,53 +6,112 @@ from mli import AsyncMLIClient
 ENDPOINT = 'http://127.0.0.1:5000'
 
 
-async def async_demo():
+async def async_demo_candle_stable_lm():
     async_client = AsyncMLIClient(ENDPOINT)
 
-    async for chunk in async_client.iter_text(**{
-        'engine': 'candle',
-        'kind': 'stable-lm',
-        'model_id': 'lmz/candle-stablelm-3b-4e1t',
-        'sample_len': 512,
-        'prompt': 'Building a website can be done in 10 simple steps:\nStep 1:'
-    }):
+    async for chunk in async_client.iter_text(
+        engine='candle',
+        kind='stable-lm',
+        model_id='lmz/candle-stablelm-3b-4e1t',
+        sample_len=512,
+        quantized=True,
+        prompt='Building a website can be done in 10 simple steps:\nStep 1:'
+    ):
         print(chunk, sep='', end='', flush=True)
 
-    async for chunk in async_client.iter_chat(**{
-        'engine': 'candle',
-        'kind': 'stable-lm',
-        'model_id': 'lmz/candle-stablelm-3b-4e1t',
-        'sample_len': 512,
-        'messages': [
+    async for chunk in async_client.iter_chat(
+        engine='candle',
+        kind='stable-lm',
+        model_id='lmz/candle-stablelm-3b-4e1t',
+        sample_len=512,
+        quantized=True,
+        messages=[
             {'role': 'system', 'content': 'You are a helpful assistant.'},
             {'role': 'user', 'content': 'I need help building a website.'},
             {'role': 'assistant', 'content': 'Sure, let me know what and hwo do you need it built.'},
-            {'role': 'user', 'content': 'Building a website can be done in 10 simple steps. Explain step by step.'}
-        ]
-    }):
+            {'role': 'user', 'content': 'Building a website can be done in 10 simple steps. Explain step by step.'},
+        ],
+    ):
         print(chunk, sep='', end='', flush=True)
 
-    print(await async_client.text(**{
-        'engine': 'candle',
-        'kind': 'stable-lm',
-        'model_id': 'lmz/candle-stablelm-3b-4e1t',
-        'sample_len': 512,
-        'prompt': 'Building a website can be done in 10 simple steps:\nStep 1:'
-    }))
+    print(await async_client.text(
+        engine='candle',
+        kind='stable-lm',
+        model_id='lmz/candle-stablelm-3b-4e1t',
+        sample_len=512,
+        prompt='Building a website can be done in 10 simple steps:\nStep 1:'
+    ))
 
-    print(await async_client.chat(**{
-        'engine': 'candle',
-        'kind': 'stable-lm',
-        'model_id': 'lmz/candle-stablelm-3b-4e1t',
-        'sample_len': 512,
-        'messages': [
+    print(await async_client.chat(
+        engine='candle',
+        kind='stable-lm',
+        model_id='lmz/candle-stablelm-3b-4e1t',
+        sample_len=512,
+        messages=[
             {'role': 'system', 'content': 'You are a helpful assistant.'},
             {'role': 'user', 'content': 'I need help building a website.'},
             {'role': 'assistant', 'content': 'Sure, let me know what and hwo do you need it built.'},
-            {'role': 'user', 'content': 'Building a website can be done in 10 simple steps. Explain step by step.'}
-        ]
-    }))
+            {'role': 'user', 'content': 'Building a website can be done in 10 simple steps. Explain step by step.'},
+        ],
+    ))
+
+
+async def async_demo_llama_cpp_main_stable_lm():
+    async_client = AsyncMLIClient(ENDPOINT)
+
+    async for chunk in async_client.iter_text(
+       engine='llama.cpp',
+        kind='main',
+        n_gpu_layers=35,
+        model_id='TheBloke/rocket-3B-GGUF',
+        model='rocket-3b.Q4_K_M.gguf',
+        sample_len=3 * 1024,
+        prompt='Building a perfect e-commerce website in 1234 simple steps:\nStep 1:',
+    ):
+        print(chunk, sep='', end='', flush=True)
+
+    async for chunk in async_client.iter_chat(
+        engine='llama.cpp',
+        kind='main',
+        n_gpu_layers=35,
+        model_id='TheBloke/rocket-3B-GGUF',
+        model='rocket-3b.Q4_K_M.gguf',
+        sample_len=3 * 1024,
+        messages=[
+            {'role': 'system', 'content': 'You are a helpful assistant.'},
+            {'role': 'user', 'content': 'I need help building a website.'},
+            {'role': 'assistant', 'content': 'Sure, let me know what and hwo do you need it built.'},
+            {'role': 'user', 'content': 'Building a perfect e-commerce website in 1234 simple steps. Explain step by step.'},
+        ],
+    ):
+        print(chunk, sep='', end='', flush=True)
+
+    print(await async_client.text(
+        engine='llama.cpp',
+        kind='main',
+        n_gpu_layers=35,
+        model_id='TheBloke/rocket-3B-GGUF',
+        model='rocket-3b.Q4_K_M.gguf',
+        sample_len=3 * 1024,
+        prompt='Building a perfect e-commerce website in 1234 simple steps:\nStep 1:',
+    ))
+
+    print(await async_client.chat(
+        engine='llama.cpp',
+        kind='main',
+        n_gpu_layers=35,
+        model_id='TheBloke/rocket-3B-GGUF',
+        model='rocket-3b.Q4_K_M.gguf',
+        sample_len=3 * 1024,
+        messages=[
+            {'role': 'system', 'content': 'You are a helpful assistant.'},
+            {'role': 'user', 'content': 'I need help building a website.'},
+            {'role': 'assistant', 'content': 'Sure, let me know what and hwo do you need it built.'},
+            {'role': 'user', 'content': 'Building a perfect e-commerce website in 1234 simple steps. Explain step by step.'},
+        ],
+    ))
 
 
 if __name__ == '__main__':
-    asyncio.run(async_demo())
+    # asyncio.run(async_demo_candle_stable_lm())
+    asyncio.run(async_demo_llama_cpp_main_stable_lm())